requirements.txt
pillow
imageio
numpy
Wand
